pyproject.toml
[tool.poetry]
name = "flask-rest-api"
version = "1.0.0"
description = "Este projeto é uma API REST desenvolvida com Flask e Python"
authors = ["Lucas Lima <bezlima.lucas@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
python-jose = "^3.3.0"
dynaconf = "^3.2.6"
psutil = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
